# services/api-smarterbot/openapi.yaml
...
openapi: 3.1.0
info:
  title: Smarter OS API
  description: Unified contact API for smarterbot.cl and smarterbot.store
  version: 1.0.0
  contact:
    email: smarterbotcl@gmail.com
  license:
    name: Proprietary
servers:
  - url: https://api.smarterbot.cl
    description: Production server
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /contact:
    post:
      summary: Submit contact form
      operationId: createContact
      tags: [contact]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactRequest'
            examples:
              basic:
                summary: Basic contact
                value:
                  name: "Juan Pérez"
                  email: "juan@example.com"
                  message: "Necesito automatizar mi facturación"
              complete:
                summary: Complete with phone and source
                value:
                  name: "María González"
                  email: "maria@company.cl"
                  message: "Quiero integrar mi CRM con Odoo"
                  phone: "+56912345678"
                  source: "smarterbot.cl"
      responses:
        '201':
          description: Contact created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail:
                  - loc: ["body", "email"]
                    msg: "value is not a valid email address"
                    type: "value_error.email"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Database not configured"
...
components:
  schemas:
    ContactRequest:
      type: object
      required:
        - name
        - email
        - message
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          description: Full name
          example: "Juan Pérez"
        email:
          type: string
          format: email
          description: Email address
          example: "juan@example.com"
        message:
          type: string
          minLength: 1
          maxLength: 5000
          description: Message content
          example: "Necesito automatizar mi facturación"
        phone:
          type: string
          maxLength: 50
          nullable: true
          description: Phone/WhatsApp number (optional)
          example: "+56912345678"
        source:
          type: string
          maxLength: 100
          nullable: true
          description: Source identifier
          example: "smarterbot.cl"
          enum:
            - smarterbot.cl
            - smarterbot.store
            - app.smarterbot.cl
    ContactResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        message:
          type: string
          example: "Contact submitted successfully"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Database not configured"
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
              msg:
                type: string
              type:
                type: string
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: JWT token for authenticated endpoints (future)

# Not applying security globally yet - will be added when authentication is implemented
# security:
#   - BearerAuth: []